Made the function recursive such that it will try to batch its attestations into smaller and smaller batches.
It will also catch singular failures, so we can be more targeted on why a batch is failing.

My main concern is what happens if say 1 out of 100 is failing the first time. On retry we may get 99 out of 100 failing because they have been previously attested. Having some sort of check would help with this. In this scenario, even if the 99 failed, and the 1 succeeded on the second try, we would be in the happy 100/100 state.