KYC is now decoupled from KYB. more rigorous testing is needed to ensure continued compliance and updates.
Instead of creating a KYB user under the LEGAL_ENTITY type, we now have every KYCUser as a USER. Legal Entities exists as their own object:
LegalEnitityController, which is the POC for a Legal Entity.

This information populates and stays to the KYCTeamId so that they can be resued in the case of organizations with multiple projects. They still get emails and populate into the Status page, but they do not need to be declared for multiple projects if selected from a selection field.